+++ zarrs/examples/array_write_read_ndarray.rs
@@ -51,7 +51,7 @@ fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
 
     println!(
         "The group metadata is:\n{}\n",
-        serde_json::to_string_pretty(&group.metadata()).unwrap()
+        group.metadata().to_string_pretty()
     );
 
     // Create an array
@@ -72,7 +72,7 @@ fn array_write_read() -> Result<(), Box<dyn std::error::Error>> {
 
     println!(
         "The array metadata is:\n{}\n",
-        serde_json::to_string_pretty(&array.metadata()).unwrap()
+        array.metadata().to_string_pretty()
     );
 
     // Write some chunks