+++ zarrs/examples/async_array_write_read.rs
@@ -47,7 +47,7 @@ async fn async_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
 
     println!(
         "The group metadata is:\n{}\n",
-        serde_json::to_string_pretty(&group.metadata())?
+        group.metadata().to_string_pretty()
     );
 
     // Create an array
@@ -68,7 +68,7 @@ async fn async_array_write_read() -> Result<(), Box<dyn std::error::Error>> {
 
     println!(
         "The array metadata is:\n{}\n",
-        serde_json::to_string_pretty(&array.metadata())?
+        array.metadata().to_string_pretty()
     );
 
     // Write some chunks