+++ zarrs/examples/custom_data_type_fixed_size.rs
@@ -274,10 +274,7 @@ fn main() {
     // .storage_transformers(vec![].into())
     .build(store, array_path)
     .unwrap();
-    println!(
-        "{}",
-        serde_json::to_string_pretty(array.metadata()).unwrap()
-    );
+    println!("{}", array.metadata().to_string_pretty());
 
     let data = [
         CustomDataTypeFixedSizeElement { x: 3, y: 4.5 },