+++ zarrs/examples/sync_http_array_read.rs
@@ -70,7 +70,7 @@ fn http_array_read(backend: Backend) -> Result<(), Box<dyn std::error::Error>> {
 
     println!(
         "The array metadata is:\n{}\n",
-        serde_json::to_string_pretty(&array.metadata()).unwrap()
+        array.metadata().to_string_pretty()
     );
 
     // Read the whole array