+++ zarrs_metadata/src/v2/array.rs
@@ -115,6 +115,13 @@ impl ArrayMetadataV2 {
         }
     }
 
+    /// Serialize the metadata as a pretty-printed String of JSON.
+    #[allow(clippy::missing_panics_doc)]
+    #[must_use]
+    pub fn to_string_pretty(&self) -> String {
+        serde_json::to_string_pretty(self).expect("array metadata is valid JSON")
+    }
+
     /// Set the dimension separator.
     #[must_use]
     pub fn with_dimension_separator(mut self, dimension_separator: ChunkKeySeparator) -> Self {